miniupnpd: updated to 2.0.20180203

+++ release/src-rt-6.x.4708/router/miniupnpd/testminissdp.c
@@ -1,8 +1,8 @@
-/* $Id: testminissdp.c,v 1.2 2017/12/12 12:01:11 nanard Exp $ */
+/* $Id: testminissdp.c,v 1.3 2018/01/15 16:46:48 nanard Exp $ */
 /* vim: tabstop=4 shiftwidth=4 noexpandtab
  * MiniUPnP project
  * http://miniupnp.free.fr/ or http://miniupnp.tuxfamily.org/
- * (c) 2017 Thomas Bernard
+ * (c) 2017-2018 Thomas Bernard
  * This software is subject to the conditions detailed
  * in the LICENCE file provided within the distribution */
 
@@ -16,7 +16,7 @@
 #include "minissdp.h"
 #include "upnpglobalvars.h"
 
-void test(const char * buffer, int n)
+void test(const char * buffer, size_t n)
 {
 	int s = 0;
 	struct sockaddr_in dummy_sender;
@@ -39,7 +39,7 @@ int main(int argc, char * * argv)
 {
 	FILE * f = NULL;
 	char buffer[1500];
-	int n = 0;
+	size_t n = 0;
 	struct lan_addr_s * lan_addr;
 
 	if((argc > 1) && ((strcmp(argv[1], "--help") == 0) || (strcmp(argv[1], "-h") == 0))) {
@@ -61,7 +61,7 @@ int main(int argc, char * * argv)
 			return 1;
 		}
 	}
-	n = (int)fread(buffer, 1, sizeof(buffer), f ? f : stdin);
+	n = fread(buffer, 1, sizeof(buffer), f ? f : stdin);
 	if(n <= 0) {
 		syslog(LOG_ERR, "error reading");
 		return 1;
